requirements.txt
Django>=5.0,<6.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.0,<5.0
django-environ>=0.11,<1.0
whitenoise>=6.6,<7.0
Pillow>=10.0,<12.0
